+++ test/clj_string_layout/test_runner.clj
@@ -2,11 +2,13 @@
   (:require [clj-string-layout.core-test]
             [clj-string-layout.escape-test]
             [clj-string-layout.property-test]
+            [clj-string-layout.table-test]
             [clojure.test :as test]))
 
 (defn -main [& _]
   (let [{:keys [fail error]} (test/run-tests 'clj-string-layout.core-test
-                                             'clj-string-layout.escape-test
-                                             'clj-string-layout.property-test)]
+                                              'clj-string-layout.escape-test
+                                              'clj-string-layout.property-test
+                                              'clj-string-layout.table-test)]
     (when (pos? (+ fail error))
       (System/exit 1))))